Move Context typedef into its sole user
