+++ packages/components/src/types.ts
@@ -9,14 +9,3 @@ export type Css = any // TODO: make this stricter to disallow { margin: 20, frui
 export type CssStatic = {}
 
 export type MessageType = "info" | "success" | "error"
-
-export interface Context {
-  pushState?: (url: string) => void
-  replaceState?: (url: string) => void
-  pushMessage: (
-    message: {
-      body: string
-      type: MessageType
-    },
-  ) => void
-}
